Add files via upload
Campus Area Network(CAN): A campus network, campus area
network, corporate area network, or CAN is a computer network made up of an
interconnection of local area networks (LANs) within a limited geographical area. The
networking types of equipment (switches, routers) and transmission media (optical fiber,
copper plant, Cat5 cabling, etc.) are almost entirely owned by the campus tenant /
owner: an enterprise, university, government, etc. A campus area network is larger
than a local area network but smaller than a metropolitan area network (MAN) or wide
area network (WAN).

NAT: Network Address Translation (NAT) is the process where a network device, usually a firewall, assigns a public address to a computer (or group of computers)
inside a private network. The main use of NAT is to limit the number of public IP
addresses an organization or company must use, for both economy and security
purposes. 

VLAN: A VLAN is a group of devices on one or more LANs that are configured
to communicate as if they were attached to the same wire, when in fact they are located
on a number of different LAN segments. Because VLANs are based on logical instead of
physical connections, they are extremely flexible. VLANs define broadcast domains in a Layer 2 network. A broadcast domain is the set of
all devices that will receive broadcast frames originating from any device within the set.